# Avoid reading id twice in AR::Core#hash